채킹 기능 초기화
